requirements.txt
streamlit==1.37.0
ultralytics==8.3.0
pillow==10.4.0
opencv-python==4.10.0.84
numpy==2.1.3
google-generativeai==0.8.3
python-dotenv==1.0.1
